mock/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
docopt==0.6.2
Flask==1.1.1
hbmqtt==0.9.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
paho-mqtt==1.4.0
passlib==1.7.1
pycryptodome==3.8.2
PyYAML==5.1
requests==2.22.0
six==1.12.0
transitions==0.6.9
urllib3==1.25.3
websockets==7.0
Werkzeug==0.15.5
